requirements.txt
Flask==2.3.3
# Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy == 3.0.0
# SQLAlchemy==1.4.46
SQLAlchemy >= 2.0, <3.0
PyMySQL==1.0.3
python-dotenv==1.0.0
Werkzeug>=2.3.7
